Recipes can now be saved and loaded per device
Currently a placeholder recipe book is in main.js and the add new recipe button isn't ready until sahana's stuff is in, but if there's no saved data, it will load the placeholder! If there is data, it will load the data saved. So, if you change the placeholder info after opening the recipes.html, it won't show because it's showing the original placeholder recipe book that's been stored into your machine